+++ cmd/vault/lock.go
@@ -2,8 +2,6 @@ package main
 
 import vaultlock "github.com/olelbis/myminivault/internal/lock"
 
-const vaultLockFile = vaultlock.DefaultFile
-
 func withVaultLock(fn func() error) error {
 	return vaultlock.WithFile(vaultLockFile, fn)
 }